requirements.txt
DAWG-Python==0.7.1
Django==1.6
PyYAML==3.11
South==1.0
argparse==1.2.1
docopt==0.6.2
pymorphy2==0.8
pymorphy2-dicts==2.4.390857.3662349
wsgiref==0.1.2
